Fix certmanager certificate status alert
Apparently cert-manager has a very strange way of reporting a certificate status to prometheus. For each certificate it reports the same metric with three different "condition" values (True, False and Unknown). If I understood it correctly, a certificate is ready when the metric with the condition label set to True is 1.

References
https://github.com/jetstack/cert-manager/blob/master/pkg/metrics/certificates.go#L77
https://github.com/jetstack/cert-manager/blob/46eaf3d1a4b5e977b612393300f0f11978ebf72e/pkg/metrics/certificates_test.go#L65-L67