Fix VPA lease checking (#155)
Apparently the VPA admission controller creates a Lease, which is checked by the updater. If the lease is not valid, the updater won't evict any pods.

By default the Lease will be created in the `kube-system` NS, unless the NAMESPACE environment variable is provided. We were setting this env variable in the admission controller Pod, but not in the updater, which meant that the Lease was being correctly created in the infrastrucutre NS, but the updater was still looking for it in the kube-system NS.